# openapi_nip_full_20251022_021407.yaml
openapi: 3.1.0
info:
  title: MEEP Clone API
  version: 3.0.0
  description: 'Patch Plan — Primeira leva (Auth, Empresas, Usuários, Eventos, Clientes,
    PDV, Ingressos, Financeiro). Padrões aplicados: RESTful, JWT bearer, paginação,
    erros tipados, idempotência em operações financeiras, webhooks como POST, versionamento
    /api/v1.

    '
servers:
- url: https://api.seudominio.com/api/v1
security:
- bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    ErrorResponse:
      description: Erro padrão
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: object
                properties:
                  code:
                    type: string
                  message:
                    type: string
                  details:
                    type: object
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
    PageSizeParam:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
    SortParam:
      name: sort
      in: query
      schema:
        type: string
    QueryParam:
      name: q
      in: query
      schema:
        type: string
...
paths:
...
  /estoque/posicao:
    get:
      tags:
      - Estoque
      summary: Listar estoque/posicao
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/PageSizeParam'
      - $ref: '#/components/parameters/SortParam'
      - $ref: '#/components/parameters/QueryParam'
      responses:
        '200':
          description: OK
        default:
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags:
      - Estoque
      summary: Criar estoque/posicao
      responses:
        '201':
          description: Created
        default:
          $ref: '#/components/responses/ErrorResponse'
...